requirements.txt
hiredis==3.0.0
packaging==24.2
python-can==4.4.2
pywin32==308
redis==5.2.0
typing_extensions==4.12.2
wrapt==1.16.0
